requirements.txt
fastapi==0.115.12
pydantic==2.11.4
pymongo==4.12.1
python_bcrypt==0.3.2
starlette==0.46.2
uvicorn